python-projects/04-rag-knowledge-assistant/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
chromadb>=0.4.22
sentence-transformers>=2.3.1
pypdf2>=3.0.1
